Source: https://github.com/msupply-foundation/open-msupply/commit/28a793b9d3ab00956d9d75efdd4f96d2bce65fd5

+++ client/packages/host/src/api/hooks/settings/namePropertyData.ts
@@ -115,7 +115,7 @@ function getForecastingPropertiesForLanguage(language: string) {
       propertyId: BUFFER_STOCK_KEY,
       key: BUFFER_STOCK_KEY,
       name: translations['BUFFER_STOCK_KEY'],
-      valueType: PropertyNodeValueType.Integer,
+      valueType: PropertyNodeValueType.Float,
       allowedValues: null,
       remoteEditable: false,
     },
